Factor out redundant definitions of IdentityRelatedField
