+++ api/v2/serializers/fields/__init__.py
@@ -3,6 +3,8 @@ from .provider_machine import ProviderMachineRelatedField
 from .user import UserRelatedField
 from .tag import TagRelatedField
 from .base import ModelRelatedField
+from .identity import IdentityRelatedField
 
 __all__ = ("ImageVersionRelatedField", "ProviderMachineRelatedField",
-           "UserRelatedField", "ModelRelatedField", "TagRelatedField")
+           "UserRelatedField", "ModelRelatedField", "TagRelatedField",
+           "IdentityRelatedField")
